v17.2.3
- Added a button to the player controls to advance to the next video in series and films.
- Fixed a bug on PC version that was resetting volume when the player recovered from a transmission error.
- Optimized search functionality to consume less resources.
- More informative error message when there is an error opening a stream.
- Prevents the scroll from being reset in search results when opening a broadcast.
- Display country of current broadcasting server by clicking the three-dot icon in the player and selecting More options > Learn more.
- Fixed PIP mode on Android.
- Improved M3U lists parsing performance.
- Better temp files auto-cleaning.
- Some minor bugfixes, improvements and code refactoring.

+++ www/nodejs-project/modules/multi-worker/multi-worker.js
@@ -118,8 +118,13 @@ const setupConstructor = () => {
 		terminate(){
 			this.finished = true
 			this.configChangeListener && global.config.removeListener('change', this.configChangeListener)
-			this.worker.terminate()
-			this.worker = null
+			if(this.worker){
+				setTimeout(() => { // try to prevent closing abruptely due to bug in v8 (FATAL ERROR: v8::FromJust Maybe value is Nothing)
+					const maybePromise = this.worker.terminate()
+					maybePromise && maybePromise.catch && maybePromise.catch(console.error)
+					this.worker = null
+				}, 3000)
+			}
 			this.rejectAll(null, 'worker manually terminated')
 			this.removeAllListeners()
 			global.config.removeListener('change', this.configChangeListener)
@@ -130,9 +135,7 @@ const setupConstructor = () => {
 			super()
 			this.Worker = require('worker_threads').Worker
 			this.worker = new this.Worker(path.join(__dirname, 'worker.js'), {
-				workerData,
-				stdout: true, 
-				stderr: true
+				workerData // leave stdout/stderr undefined
 			})
 			this.worker.on('error', err => {
 				let serr = String(err)
@@ -161,7 +164,7 @@ const setupConstructor = () => {
 						this.promises[ret.id][ret.type](ret.data)
 						delete this.promises[ret.id]
 					} else {
-						console.warn('Callback repeated: '+ JSON.stringify(ret))
+						console.warn('Callback repeated', ret)
 					}
 				} else {
 					let args = [], pos = ret.data.indexOf(':')
@@ -170,6 +173,9 @@ const setupConstructor = () => {
 						let evtContent = ret.data.substr(pos + 1)
 						if(evtContent.length){
 							evtContent = global.parseJSON(evtContent)
+						}
+						if(typeof(evtContent) == 'object' && evtContent.type == 'Buffer') {
+							evtContent = Buffer.from(evtContent.data)
 						}
 						args = [evtType, evtContent]
 					} else {
